refactor(commands): extract audio, ask, meetings domains (gated/masked, verbatim)
Pure MOVE, zero behavior change. Three gate/mask-touching command domains
lift out of commands/mod.rs into commands/{audio,ask,meetings}.rs (each
#[path]-aliased — the names shadow crate::{audio}); mod.rs -1373 lines
(35,955 -> 34,582), lib.rs BYTE-FOR-BYTE UNCHANGED.

All 63 moved #[tauri::command]s + helpers are BYTE-IDENTICAL to trunk
(verbatim; only fn->pub(crate) widening where a staying command/test
calls via the re-export). Gate/mask/seal logic UNCHANGED, only relocated:
- audio: list_voiceprints (excludes sealed voiceprints), free_up_space
  (holds lifecycle_guard across the prune), the recording/voice-command
  commands. No new at-rest path, no on-disk audio path to the FE.
- ask: the agentic-Ask loop + list_assistant_threads (fails closed to
  empty unlock set) + gated_meeting_thread_turns (gated corpus).
- meetings: get_meeting_detail (-> meeting_is_unlocked -> masked_detail:
  locked:true / no note/segments / audio_path:None), get_timeline (empty
  when sealed), list_meetings (mask_locked_meetings), search_meetings.
The shared mask/gate helpers masked_detail + mask_locked_meetings stay in
mod.rs, promoted pub(crate) (bodies byte-identical); meeting_is_unlocked/
unlocked_snapshot already pub(crate).

Tests stay in mod.rs (#[test] 346 = trunk; lifecycle_tests 313/0 core
lock/seal/mask suite green). Verified: cargo test --lib audio/voiceprint/
ask/assistant/meeting/timeline/speaker/tag + lock_read_gate all green,
clippy --lib -D warnings clean, MSRV grep clean.
